refactor(saves): extract pure helpers to saves/_helpers.py
Move _compute_uploaded_by_us (uploader attribution) and
_local_save_target (canonical filename) out of SaveService into
a private _helpers module. Both are static and used cross-module
by upcoming sub-services. _local_save_target loses its
@staticmethod decorator since it's no longer on a class.

Refs #257.
